tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssimportrule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSImportRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssimportrule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSImportRule' in window);

      }, '\'CSSImportRule\' in window');

      test(function(){

        assert_true(CSSImportRule instanceof Function);

      }, 'CSSImportRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSImportRule);

      }, '\'prototype\' in CSSImportRule');

      test(function(){

        assert_equals(CSSImportRule.prototype.toString(), '[object CSSImportRulePrototype]');

      }, 'CSSImportRule.prototype.toString() == \'[object CSSImportRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSImportRule.prototype);

      }, '\'constructor\' in CSSImportRule.prototype');

      test(function(){

        assert_equals(CSSImportRule.prototype.constructor,CSSImportRule);

      }, 'CSSImportRule.prototype.constructor === CSSImportRule');

    ]]>
    </script>
  </body>
</html>
